requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
anthropic==0.28.1
supabase==2.4.2
Gunicorn==21.2.0
